requirements.txt
# LLM local
transformers>=4.38.0
ollama>=llama3.1:8b
accelerate>=0.20.0

# Imagens
diffusers>=0.25.0
safetensors>=0.4.0
Pillow>=10.0.0

# TTS
coqui-tts>=0.15.0

# Vídeo
moviepy>=1.0.3
pydub>=0.25.1

# YouTube upload
google-api-python-client>=2.70.0
oauth2client>=4.1.3

# Outros
pyyaml>=6.0
requests>=2.31.0
python-dotenv>=1.0.0
